pyproject.toml
[project]
name = "ortobahn"
version = "0.1.0"
description = "Autonomous AI marketing engine - a Vaultscaler subsidiary"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "atproto>=0.0.55",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "feedparser>=6.0",
    "newsapi-python>=0.2.7",
    "pytrends>=4.9",
    "tweepy>=4.14.0",
    "requests>=2.31.0",
]

[project.scripts]
ortobahn = "ortobahn.__main__:main"

[project.optional-dependencies]
web = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1",
    "email-validator>=2.0",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "ruff>=0.4",
    "mypy>=1.8",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
include = ["ortobahn*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=ortobahn --cov-report=term-missing -v"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
disable_error_code = ["override"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
